+++ plugin/__init__.py
@@ -1,3 +1,3 @@
 """EDPPMT: Elite Dangerous PowerPlay Merit Tracker."""
 
-__version__ = "1.7.3"
+__version__ = "1.8.0"